requirements.txt
Flask==2.0.1
spotipy==2.19.0
pandas==1.4.1
scikit-learn==1.0.2
python-dotenv==0.20.0